+++ CCHMapClusterController Tests/CCHMapTreeTests.m
@@ -87,7 +87,7 @@
     MKPointAnnotation *annotation0 = [[MKPointAnnotation alloc] init];
     annotation0.coordinate = CLLocationCoordinate2DMake(52, 13);
     [self.mapTree addAnnotations:@[annotation0, annotation0]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
 }
 
 - (void)testAnnotationsInMapRectContainsRetain
@@ -101,12 +101,12 @@
         
         [self.mapTree addAnnotations:@[annotation]];
     }
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
 
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)1);
+    XCTAssertEqual(annotations.count, 1);
     if (annotations.count > 0) {
         XCTAssertEqualObjects(title, [annotations.anyObject title]);
     }
@@ -114,12 +114,12 @@
 
 - (void)testAnnotationsInMapRectEmpty
 {
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)0);
+    XCTAssertEqual(self.mapTree.annotations.count, 0);
 
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(CLLocationCoordinate2DMake(52, 13), 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)0);
+    XCTAssertEqual(annotations.count, 0);
 }
 
 - (void)testAnnotationsInMapRectContains
@@ -127,12 +127,12 @@
     MKPointAnnotation *annotation = [[MKPointAnnotation alloc] init];
     annotation.coordinate = CLLocationCoordinate2DMake(52, 13);
     [self.mapTree addAnnotations:@[annotation]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(annotation.coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)1);
+    XCTAssertEqual(annotations.count, 1);
     if (annotations.count > 0) {
         XCTAssertEqual(annotation, annotations.anyObject);
     }
@@ -150,7 +150,7 @@
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(annotation0.coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)2);
+    XCTAssertEqual(annotations.count, 2);
     if (annotations.count > 0) {
         XCTAssertTrue([annotations containsObject:annotation0]);
         XCTAssertTrue([annotations containsObject:annotation1]);
@@ -162,13 +162,13 @@
     MKPointAnnotation *annotation = [[MKPointAnnotation alloc] init];
     annotation.coordinate = CLLocationCoordinate2DMake(52, 13);
     [self.mapTree addAnnotations:@[annotation]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     
     CLLocationCoordinate2D coordinate = CLLocationCoordinate2DMake(50, 10);
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)0);
+    XCTAssertEqual(annotations.count, 0);
 }
 
 - (void)testAnnotationsInMapRectContainsSome
@@ -178,12 +178,12 @@
     MKPointAnnotation *annotation1 = [[MKPointAnnotation alloc] init];
     annotation1.coordinate = CLLocationCoordinate2DMake(50, 10);
     [self.mapTree addAnnotations:@[annotation0, annotation1]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)2);
+    XCTAssertEqual(self.mapTree.annotations.count, 2);
     
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(annotation1.coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)1);
+    XCTAssertEqual(annotations.count, 1);
     if (annotations.count > 0) {
         XCTAssertEqual(annotation1, annotations.anyObject);
     }
@@ -198,31 +198,31 @@
     MKPointAnnotation *annotation2 = [[MKPointAnnotation alloc] init];
     annotation1.coordinate = annotation1.coordinate;
     [self.mapTree addAnnotations:@[annotation0, annotation1, annotation2]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)3);
+    XCTAssertEqual(self.mapTree.annotations.count, 3);
 
     [self.mapTree removeAnnotations:@[annotation0]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)2);
+    XCTAssertEqual(self.mapTree.annotations.count, 2);
     MKCoordinateRegion region = MKCoordinateRegionMakeWithDistance(annotation0.coordinate, 1000, 1000);
     MKMapRect mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     NSSet *annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)0);
+    XCTAssertEqual(annotations.count, 0);
 
     [self.mapTree removeAnnotations:@[annotation2]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     region = MKCoordinateRegionMakeWithDistance(annotation1.coordinate, 1000, 1000);
     mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)1);
+    XCTAssertEqual(annotations.count, 1);
     if (annotations.count > 0) {
         XCTAssertEqual(annotation1, annotations.anyObject);
     }
 
     [self.mapTree removeAnnotations:@[annotation1]];
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)0);
+    XCTAssertEqual(self.mapTree.annotations.count, 0);
     region = MKCoordinateRegionMakeWithDistance(annotation1.coordinate, 1000, 1000);
     mapRect = CCHMapClusterControllerMapRectForCoordinateRegion(region);
     annotations = [self.mapTree annotationsInMapRect:mapRect];
-    XCTAssertEqual(annotations.count, (NSUInteger)0);
+    XCTAssertEqual(annotations.count, 0);
 }
 
 - (void)testAddRemoveAnnotationsUpdated
@@ -232,29 +232,29 @@
     annotation0.id = @"123";
     BOOL updated = [self.mapTree addAnnotations:@[annotation0]];
     XCTAssertTrue(updated);
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     
     // Add again
     updated = [self.mapTree addAnnotations:@[annotation0]];
     XCTAssertFalse(updated);
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     
     // Add equal
     Annotation *annotation1 = [[Annotation alloc] init];
     annotation1.id = annotation0.id;
     updated = [self.mapTree addAnnotations:@[annotation1]];
     XCTAssertFalse(updated);
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)1);
+    XCTAssertEqual(self.mapTree.annotations.count, 1);
     
     // Remove equal
     updated = [self.mapTree removeAnnotations:@[annotation1]];
     XCTAssertTrue(updated);
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)0);
+    XCTAssertEqual(self.mapTree.annotations.count, 0);
     
     // Remove again
     updated = [self.mapTree removeAnnotations:@[annotation0]];
     XCTAssertFalse(updated);
-    XCTAssertEqual(self.mapTree.annotations.count, (NSUInteger)0);
+    XCTAssertEqual(self.mapTree.annotations.count, 0);
 }
 
 @end
